Project 2: Short Text Location Prediction
The goal of this Project is to build and critically analyse some supervised Machine Learning methods, with the aim of automatically identifying the location from which a textual message was sent, as one of four Australian cities. Although this is a simplification of the more general problem of geotagging, it is still a very difficult problem, which has been well-studied, but a solution remains elusive. This aims to reinforce the largely theoretical lecture concepts surrounding learners, data representation, and evaluation, by applying them to a sophisticated problem. You will also have an opportunity to practice your general problem-solving skills, written communication skills, and creativity.